[IMP] Moved case notes socketIO event handlers to new namespace app.blueprints.socket_io_event_handlers
